+++ app/src/main/java/edu/uiuc/anymap/Point.java
@@ -4,6 +4,6 @@ package edu.uiuc.anymap;
  * Created by KomIG on 9/13/14.
  */
 public class Point {
-    public static float x;
-    public static float y;
+    public static double x;
+    public static double y;
 }
